requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
stripe>=7.0.0
APScheduler>=3.10.0 